feat: refactor LoginPage as orchestrator with extracted Auth components
- Extract LoginForm component from monolithic LoginPage
- Extract RegistrationForm component from monolithic LoginPage
- Fix LoginForm interface to return complete AuthResponse instead of just token
- Fix RegistrationForm interface to return complete AuthResponse instead of void
- Create lightweight LoginPage orchestrator that delegates to components
- Add comprehensive unit tests for both components (11 total tests)
- Follow src/components/Auth/** directory structure
- Maintain session management and redirect functionality
